Extract flavor switcher action menu
